Working Pimpl-Version of Index
The Index class is rather expensive to compile and needs Access to many datatypes that aren't needed elsewhere.
This commit restructures the Index class to use the Pimpl (Pointer to implementation) idiom, s.t. we can change
the Index implementation without retriggering the expensive compilation of all different Operations.
