+++ src/WriteIndexListsMain.cpp
@@ -9,6 +9,7 @@
 #include <libgen.h>
 #include <stdlib.h>
 
+#include <fstream>
 #include <iomanip>
 #include <iostream>
 #include <string>
