+++ src/engine/Server.h
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <nlohmann/json.hpp>
 #include <string>
 #include <vector>
 
@@ -22,6 +23,8 @@ using std::string;
 using std::vector;
 
 using ad_utility::Socket;
+
+using nlohmann::json;
 
 //! The HTTP Sever used.
 class Server {
