+++ src/index/CMakeLists.txt
@@ -1,7 +1,8 @@
 
 add_library(index
-        Index.h Index.cpp Index.Text.cpp
-        Vocabulary.h VocabularyImpl.h
+        Index.h Index.cpp
+        IndexImpl.h IndexImpl.cpp IndexImpl.Text.cpp
+        Vocabulary.h Vocabulary.cpp
         VocabularyGenerator.h VocabularyGeneratorImpl.h
         ConstantsIndexCreation.h
         ExternalVocabulary.h ExternalVocabulary.cpp
