+++ src/index/IndexBuilderMain.cpp
@@ -3,6 +3,7 @@
 // Author: Björn Buchhold (buchhold@informatik.uni-freiburg.de)
 
 #include <getopt.h>
+
 #include <cstdio>
 #include <cstdlib>
 #include <exception>
@@ -10,9 +11,12 @@
 #include <iostream>
 #include <sstream>
 #include <string>
-
 #include <unordered_set>
+
 #include "../global/Constants.h"
+#include "../parser/NTriplesParser.h"
+#include "../parser/TsvParser.h"
+#include "../parser/TurtleParser.h"
 #include "../util/File.h"
 #include "../util/ReadableNumberFact.h"
 #include "../util/StringUtils.h"
