+++ test/GroupByTest.cpp
@@ -3,9 +3,12 @@
 // Author: Florian Kramer (florian.kramer@mail.uni-freiburg.de)
 
 #include <gtest/gtest.h>
+
 #include <cstdio>
+
 #include "../src/engine/CallFixedSize.h"
 #include "../src/engine/GroupBy.h"
+#include "../src/parser/NTriplesParser.h"
 
 ad_utility::AllocatorWithLimit<Id>& allocator() {
   static ad_utility::AllocatorWithLimit<Id> a{
